Removed DiscreteStochastic and BinaryStochastic, added dtype argument to Variable.
git-svn-id: https://pymc.googlecode.com/svn/trunk@631 15d7aa0b-6f1a-0410-991a-d59f85d14984

+++ pymc/examples/gelman_bioassay.py
@@ -18,7 +18,7 @@ def theta(a=alpha, b=beta, d=dose):
     return invlogit(a+b*d)
 
 @data
-@stochastic
+@stochastic(dtype=int)
 def deaths(value=array([0,1,3,5],dtype=float), n=n, p=theta):
     """deaths ~ binomial(n, p)"""
     return binomial_like(value, n, p)
